# LVStyleSheet: optimize `LVStyleSheet::merge`