@benoit-pierre 2023-02-19
> Reviewed 2 of 2 files at r1, all commit messages.
> _[Reviewable](https://reviewable.io/reviews/koreader/crengine/509)_ status: :shipit: complete! all files reviewed, all discussions resolved (waiting on @benoit-pierre)

Can you explain that part of the workflow? Why is the status `waiting on me`? What do I need to do?