@NiLuJe 2023-02-19
That's just because you're the PR's author ;).

c.f., https://github.com/koreader/koreader-base/pull/1584 e.g., where it's waiting on... myself ;).

It's slightly confusing to be flagged as waiting when no comments were made, but it's basically to make the PR author aware that a review was done and may require ACK/NACK of comments and/or code changes.